+++ photon-model-tasks/src/main/java/com/vmware/photon/controller/model/tasks/PhotonModelTaskServices.java
@@ -14,9 +14,11 @@
 package com.vmware.photon.controller.model.tasks;
 
 import com.vmware.photon.controller.model.tasks.monitoring.ResourceStatsAggregationTaskService;
+import com.vmware.photon.controller.model.tasks.monitoring.SingleResourceStatsAggregationTaskService;
 import com.vmware.photon.controller.model.tasks.monitoring.SingleResourceStatsCollectionTaskService;
 import com.vmware.photon.controller.model.tasks.monitoring.StatsAggregationTaskService;
 import com.vmware.photon.controller.model.tasks.monitoring.StatsCollectionTaskService;
+
 import com.vmware.xenon.common.Operation;
 import com.vmware.xenon.common.ServiceHost;
 import com.vmware.xenon.services.common.TaskFactoryService;
@@ -40,7 +42,8 @@ public class PhotonModelTaskServices {
             SingleResourceStatsCollectionTaskService.FACTORY_LINK,
             StatsAggregationTaskService.FACTORY_LINK,
             EndpointAllocationTaskService.FACTORY_LINK,
-            ResourceStatsAggregationTaskService.FACTORY_LINK };
+            ResourceStatsAggregationTaskService.FACTORY_LINK,
+            SingleResourceStatsAggregationTaskService.FACTORY_LINK };
 
     public static void startServices(ServiceHost host) throws Throwable {
 
@@ -75,5 +78,7 @@ public class PhotonModelTaskServices {
                 () -> TaskFactoryService.create(EndpointRemovalTaskService.class));
         host.startFactory(ResourceStatsAggregationTaskService.class,
                 () -> TaskFactoryService.create(ResourceStatsAggregationTaskService.class));
+        host.startFactory(SingleResourceStatsAggregationTaskService.class,
+                () -> TaskFactoryService.create(SingleResourceStatsAggregationTaskService.class));
     }
 }