+++ photon-model-tasks/src/main/java/com/vmware/photon/controller/model/tasks/monitoring/StatsConstants.java
@@ -30,7 +30,7 @@ public class StatsConstants {
     public static final int NUM_BUCKETS_DAILY_DATA = 4 * 7;
 
     // size of the bucket in milliseconds for maintaining data at the granularity of an hour
-    public static final int BUCKET_SIZE_HOURS_IN_MILLS = BUCKET_SIZE_MINUTES_IN_MILLIS * 60;
+    public static final int BUCKET_SIZE_HOURS_IN_MILLIS = BUCKET_SIZE_MINUTES_IN_MILLIS * 60;
     // size of the bucket in milliseconds for maintaining data at the granularity of a day
-    public static final int BUCKET_SIZE_DAYS_IN_MILLS = BUCKET_SIZE_HOURS_IN_MILLS * 24;
+    public static final int BUCKET_SIZE_DAYS_IN_MILLIS = BUCKET_SIZE_HOURS_IN_MILLIS * 24;
 }